+++ responses/catalogData.json
@@ -1,41 +1,41 @@
 [
   {
-    "id_product": 123,
-    "product_name": "Refrigerator",
+    "id": 123,
+    "title": "Refrigerator",
     "price": 45600
   },
   {
-    "id_product": 456,
-    "product_name": "Keyboard",
+    "id": 456,
+    "title": "Keyboard",
     "price": 2000
   },
   {
-    "id_product": 345,
-    "product_name": "Phon",
+    "id": 345,
+    "title": "Phon",
     "price": 45600
   },
   {
-    "id_product": 898,
-    "product_name": "Computer",
+    "id": 898,
+    "title": "Computer",
     "price": 1000
   },
   {
-    "id_product": 457,
-    "product_name": "Memmory",
+    "id": 457,
+    "title": "Memmory",
     "price": 45600
   },
   {
-    "id_product": 232,
-    "product_name": "Netbook",
+    "id": 232,
+    "title": "Netbook",
     "price": 23657
   },
   {
-    "id_product": 555,
-    "product_name": "Monitor",
+    "id": 555,
+    "title": "Monitor",
     "price": 33000
   },
   {
-    "id_product": 456,
+    "id": 456,
     "product_name": "Cooker",
     "price": 555000
   }
