+++ src/index.html
@@ -1,0 +1,76 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <title>Интернет-магазин</title>
+   <!--  <link rel="stylesheet" href="./src/css/style.css"> -->
+</head>
+
+<body>
+
+
+    <div class="cart">
+
+        <div class="cart__body">
+
+            <span class="cart__close">x</span>
+
+            <h1>Cart</h1>
+
+            <hr>
+
+            <div class="cart__headerBody">
+
+                <p class="colImage">Image</p>
+                <p class="colTitle">Title</p>
+                <p class="colPrice">Price</p>
+                <p class="colAllPrice">allPrice</p>
+                <p class="colQuantity">Quantity</p>
+
+            </div>
+
+            <div class="cart__products">
+
+
+            </div>
+
+            <hr>
+
+            <div class="cart__resalts">
+
+                <p class="cart__allQuantity">
+                    <strong>All quantity </strong>
+
+                    <span id="allquantity"> 0 </span>
+                </p>
+
+                <p class="cart__allSum">
+                    <strong>All sum </strong>
+
+                    <span id="allsum"> 0 </span>
+                </p>
+
+                <button id="claerCart">Clear cart</button>
+
+            </div>
+
+        </div>
+
+    </div>
+    <div class="wrapper">
+        <header>
+            <a href="index.html" class="logo"> Js2 </a>
+            <button class="btn-cart" type="button">Корзина</button>
+        </header>
+
+        <main>
+            <div class="products"></div>
+        </main>
+
+    </div>
+
+ 
+</body>
+
+</html>
