+++ src/js/constants.js
@@ -1,0 +1,5 @@
+const constants = {
+    baseURL : 'https://raw.githubusercontent.com/lotostoi/js2GeekBrains/lesson3/responses/'
+}
+
+export default constants
